refactore : 글 등록 페이지 코드 분리
